refactor(socket): self-register commands in registry
