+++ src/main/java/io/taanielo/jmud/core/server/socket/RegistrableCommand.java
@@ -1,0 +1,12 @@
+package io.taanielo.jmud.core.server.socket;
+
+import java.util.Objects;
+
+/**
+ * Base command that registers itself with the provided registry.
+ */
+public abstract class RegistrableCommand implements SocketCommandHandler {
+    protected RegistrableCommand(SocketCommandRegistry registry) {
+        Objects.requireNonNull(registry, "Command registry is required").register(this);
+    }
+}
